+++ Tests/DatasourceTestCase.swift
@@ -22,8 +22,6 @@ class DatasourceTestCase: XCTestCase {
         controller = DTTestTableViewController()
         controller.tableView = AlwaysVisibleTableView()
         let _ = controller.view
-        controller.manager.startManaging(withDelegate: controller)
-        controller.manager.storage = MemoryStorage()
         
         controller.manager.register(NibCell.self)
     }