+++ Tests/MappingConditionsTestCase.swift
@@ -20,8 +20,6 @@ class MappingConditionsTestCase: XCTestCase {
         controller = DTTestTableViewController()
         controller.tableView = AlwaysVisibleTableView()
         let _ = controller.view
-        controller.manager.startManaging(withDelegate: controller)
-        controller.manager.storage = MemoryStorage()
     }
     
     func testMappingCanBeSwitchedBetweenSections() {