+++ Tests/MappingTestCase.swift
@@ -21,8 +21,6 @@ class MappingTestCase: XCTestCase {
         controller = DTTestTableViewController()
         controller.tableView = AlwaysVisibleTableView()
         let _ = controller.view
-        controller.manager.startManaging(withDelegate: controller)
-        controller.manager.storage = MemoryStorage()
     }
 
     func testShouldCreateCellFromCode()